Bug 29844 follow-up (as_list)
Bug 29844 removed uses of wantarray in Koha::Objects. So, to get a lists that can be iterated in the report-step1.tt template a call to as_list is needed.  Otherwise the script fails with error messages:
The method Koha::Libraries->branchcode is not covered by tests
and
The method Koha::Patron::Categories->categorycode is not covered by tests